fix a compiler warning

+++ src/modulefinder/sentry_modulefinder_linux.c
@@ -313,7 +313,8 @@ sentry__procmaps_module_to_value(const sentry_module_t *module)
     sentry_value_set_by_key(
         mod_val, "image_addr", sentry__value_new_addr((uint64_t)module->start));
     sentry_value_set_by_key(mod_val, "image_size",
-        sentry_value_new_int32((int32_t)(module->end - module->start)));
+        sentry_value_new_int32(
+            (int32_t)((size_t)module->end - (size_t)module->start)));
     sentry_value_set_by_key(mod_val, "code_file",
         sentry__value_new_string_owned(sentry__slice_to_owned(module->file)));
